+++ backend/app/models/client.py
@@ -2,6 +2,7 @@
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import Boolean, DateTime, ForeignKey, Index, Numeric, String, func, text
 from sqlalchemy.dialects.postgresql import UUID
@@ -31,10 +32,10 @@ class Client(Base):
     functional_currency: Mapped[str] = mapped_column(
         String(3), nullable=False, server_default=text("'GBP'")
     )
-    materiality_threshold_pct: Mapped[float] = mapped_column(
+    materiality_threshold_pct: Mapped[Decimal] = mapped_column(
         Numeric(5, 2), nullable=False, server_default=text("10.00")
     )
-    materiality_threshold_abs: Mapped[float] = mapped_column(
+    materiality_threshold_abs: Mapped[Decimal] = mapped_column(
         Numeric(19, 2), nullable=False, server_default=text("1000.00")
     )
     is_deleted: Mapped[bool] = mapped_column(
